refactor(runner): split local submit tests by responsibility
