+++ crates/runner/src/cmd/local/submit.rs
@@ -742,1748 +742,4 @@ async fn run_submit_with_home(args: SubmitArgs, home: HomePaths) -> RunnerResult
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::test_fixtures::{ignored_child_test_env_guard_enabled, run_ignored_child_test};
-    use std::os::unix::fs::{PermissionsExt, symlink};
-
-    const TIMEZONE_CHILD_SCENARIO: &str = "VM0_RUNNER_TIMEZONE_TEST_SCENARIO";
-    const TIMEZONE_CHILD_TEST: &str = "cmd::local::submit::tests::detect_system_timezone_child";
-    const TIMEZONE_FROM_ENV_SCENARIO: &str = "from-env";
-    const TIMEZONE_EMPTY_ENV_SCENARIO: &str = "empty-env";
-    const TEST_QUEUE_WATCH_TIMEOUT: Duration = Duration::from_secs(5);
-    const TEST_QUEUE_WATCH_INTERVAL: Duration = Duration::from_millis(1);
-
-    fn submit_queue_entry(group_dir: &Path, job_id: RunId) -> SubmitQueueEntry {
-        SubmitQueueEntry::for_job(group_dir, crate::profile::DEFAULT_PROFILE, job_id).unwrap()
-    }
-
-    fn write_queue_job_file(group_dir: &Path, profile: &str, job_id: RunId) -> PathBuf {
-        let path = local_queue::job_path(group_dir, profile, job_id).unwrap();
-        std::fs::create_dir_all(path.parent().unwrap()).unwrap();
-        std::fs::write(&path, b"{}").unwrap();
-        path
-    }
-
-    fn mode(path: &Path) -> u32 {
-        std::fs::metadata(path).unwrap().permissions().mode() & 0o777
-    }
-
-    #[tokio::test]
-    async fn detect_system_timezone_from_env() {
-        run_timezone_child(TIMEZONE_FROM_ENV_SCENARIO, "America/New_York").await;
-    }
-
-    #[tokio::test]
-    async fn detect_system_timezone_empty_env() {
-        run_timezone_child(TIMEZONE_EMPTY_ENV_SCENARIO, "").await;
-    }
-
-    async fn run_timezone_child(scenario: &'static str, timezone: &'static str) {
-        run_ignored_child_test(
-            TIMEZONE_CHILD_TEST,
-            (TIMEZONE_CHILD_SCENARIO, scenario),
-            &[("TZ", Some(timezone))],
-            Duration::from_secs(5),
-        )
-        .await;
-    }
-
-    #[test]
-    #[ignore = "spawned by timezone environment scenario tests"]
-    fn detect_system_timezone_child() {
-        let Ok(scenario) = std::env::var(TIMEZONE_CHILD_SCENARIO) else {
-            return;
-        };
-        if !ignored_child_test_env_guard_enabled((TIMEZONE_CHILD_SCENARIO, &scenario)) {
-            return;
-        }
-
-        match scenario.as_str() {
-            TIMEZONE_FROM_ENV_SCENARIO => {
-                assert_eq!(
-                    detect_system_timezone(),
-                    Some("America/New_York".to_string())
-                );
-            }
-            TIMEZONE_EMPTY_ENV_SCENARIO => {
-                // Empty TZ falls through to /etc/timezone.
-                assert_ne!(detect_system_timezone(), Some("".to_string()));
-            }
-            other => panic!("unknown timezone test scenario: {other}"),
-        }
-    }
-
-    #[test]
-    fn try_read_result_nonexistent_returns_none() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("does-not-exist.result");
-        assert!(try_read_result(&path).is_none());
-    }
-
-    #[test]
-    fn try_read_result_empty_returns_none() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("empty.result");
-        std::fs::write(&path, b"").unwrap();
-        assert!(try_read_result(&path).is_none());
-    }
-
-    #[test]
-    fn try_read_result_with_content_returns_some() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("valid.result");
-        std::fs::write(&path, b"{\"exit_code\":0}").unwrap();
-        let result = try_read_result(&path).unwrap();
-        assert_eq!(result, b"{\"exit_code\":0}");
-    }
-
-    #[test]
-    fn try_read_result_ignores_result_file_symlink() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let result_path = local_queue::result_path(group_dir, RunId::new_v4());
-        std::fs::create_dir_all(result_path.parent().unwrap()).unwrap();
-        let target = dir.path().join("target-result");
-        std::fs::write(&target, b"{\"exit_code\":0}").unwrap();
-        symlink(&target, &result_path).unwrap();
-
-        assert!(try_read_result(&result_path).is_none());
-    }
-
-    #[test]
-    fn result_file_is_empty_ignores_result_file_symlink() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let result_path = local_queue::result_path(group_dir, RunId::new_v4());
-        std::fs::create_dir_all(result_path.parent().unwrap()).unwrap();
-        let target = dir.path().join("target-result");
-        std::fs::write(&target, b"").unwrap();
-        symlink(&target, &result_path).unwrap();
-
-        assert!(!result_file_is_empty(&result_path));
-    }
-
-    #[test]
-    fn abandoned_marker_write_publishes_without_tmp_residue() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let result_path = local_queue::result_path(group_dir, job_id);
-
-        let marker =
-            write_abandoned_result_marker(&result_path, job_id, "local submit abandoned").unwrap();
-
-        assert_eq!(std::fs::read(&result_path).unwrap(), marker.bytes);
-        assert_eq!(mode(&result_path), 0o600);
-        let result_dir = local_queue::results_dir(group_dir);
-        assert_eq!(mode(&result_dir), 0o700);
-        let tmp_files: Vec<_> = std::fs::read_dir(result_dir)
-            .unwrap()
-            .filter_map(Result::ok)
-            .filter(|entry| entry.path().extension().and_then(|ext| ext.to_str()) == Some("tmp"))
-            .collect();
-        assert!(tmp_files.is_empty(), "tmp files left behind: {tmp_files:?}");
-    }
-
-    #[test]
-    fn abandoned_marker_write_creates_missing_group_dir_as_shared_trusted() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path().join("groups").join("org").join("group");
-        let job_id = RunId::new_v4();
-        let result_path = local_queue::result_path(&group_dir, job_id);
-
-        let marker =
-            write_abandoned_result_marker(&result_path, job_id, "local submit abandoned").unwrap();
-
-        assert_eq!(std::fs::read(&result_path).unwrap(), marker.bytes);
-        assert_eq!(mode(&group_dir), crate::host_file::SHARED_TRUSTED_DIR_MODE);
-        assert_eq!(mode(&local_queue::results_dir(&group_dir)), 0o700);
-        assert_eq!(mode(&result_path), 0o600);
-    }
-
-    #[test]
-    fn abandoned_marker_write_cleans_tmp_when_publish_fails() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let result_path = local_queue::result_path(group_dir, job_id);
-        std::fs::create_dir_all(&result_path).unwrap();
-
-        let marker = write_abandoned_result_marker(&result_path, job_id, "local submit abandoned");
-
-        assert!(marker.is_none());
-        assert!(result_path.is_dir());
-        let result_dir = local_queue::results_dir(group_dir);
-        let tmp_files: Vec<_> = std::fs::read_dir(result_dir)
-            .unwrap()
-            .filter_map(Result::ok)
-            .filter(|entry| entry.path().extension().and_then(|ext| ext.to_str()) == Some("tmp"))
-            .collect();
-        assert!(tmp_files.is_empty(), "tmp files left behind: {tmp_files:?}");
-    }
-
-    #[test]
-    fn abandoned_marker_write_preserves_existing_empty_result() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let result_path = local_queue::result_path(group_dir, job_id);
-        std::fs::create_dir_all(result_path.parent().unwrap()).unwrap();
-        std::fs::write(&result_path, b"").unwrap();
-
-        let marker = write_abandoned_result_marker(&result_path, job_id, "local submit abandoned");
-
-        assert!(marker.is_none());
-        assert!(result_file_is_empty(&result_path));
-        let result_dir = local_queue::results_dir(group_dir);
-        let tmp_files: Vec<_> = std::fs::read_dir(result_dir)
-            .unwrap()
-            .filter_map(Result::ok)
-            .filter(|entry| entry.path().extension().and_then(|ext| ext.to_str()) == Some("tmp"))
-            .collect();
-        assert!(tmp_files.is_empty(), "tmp files left behind: {tmp_files:?}");
-    }
-
-    #[test]
-    fn abandoned_cleanup_keeps_replaced_result_with_same_content() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.job.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.job, b"{}").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-        let marker =
-            write_abandoned_result_marker(&queue.result, job_id, "local submit abandoned").unwrap();
-        let replacement_path = queue.result.with_extension("replacement");
-        std::fs::write(&replacement_path, &marker.bytes).unwrap();
-        std::fs::rename(&replacement_path, &queue.result).unwrap();
-
-        queue.cleanup_abandoned(Some(&marker));
-
-        assert!(
-            queue.result.exists(),
-            "cleanup must not remove a result that replaced the submit marker"
-        );
-    }
-
-    #[test]
-    fn abandoned_cleanup_keeps_mutated_result_with_same_marker_inode() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.job.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.job, b"{}").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-        let marker =
-            write_abandoned_result_marker(&queue.result, job_id, "local submit abandoned").unwrap();
-        let marker_metadata = std::fs::metadata(&queue.result).unwrap();
-        let runner_result = b"runner result";
-        let mut result_file = std::fs::OpenOptions::new()
-            .write(true)
-            .truncate(true)
-            .open(&queue.result)
-            .unwrap();
-        std::io::Write::write_all(&mut result_file, runner_result).unwrap();
-        drop(result_file);
-        let current_metadata = std::fs::metadata(&queue.result).unwrap();
-        assert_eq!(marker_metadata.dev(), current_metadata.dev());
-        assert_eq!(marker_metadata.ino(), current_metadata.ino());
-
-        queue.cleanup_abandoned(Some(&marker));
-
-        assert!(!queue.job.exists());
-        assert_eq!(std::fs::read(&queue.result).unwrap(), runner_result);
-        assert!(!queue.cancel.exists());
-        assert!(!queue.claim.exists());
-    }
-
-    async fn wait_for_job_and_write_result(
-        group_dir: std::path::PathBuf,
-        profile: String,
-        exit_code: i32,
-        error: Option<String>,
-    ) -> JobRequest {
-        let job_dir = local_queue::profile_jobs_dir(&group_dir, &profile).unwrap();
-        loop {
-            if let Ok(entries) = std::fs::read_dir(&job_dir) {
-                for entry in entries.filter_map(Result::ok) {
-                    let path = entry.path();
-                    if path.extension().and_then(|ext| ext.to_str()) != Some("job") {
-                        continue;
-                    }
-                    let request: JobRequest =
-                        serde_json::from_slice(&std::fs::read(&path).unwrap()).unwrap();
-                    let response = JobResponse {
-                        run_id: request.job_id,
-                        exit_code,
-                        error: error.clone(),
-                    };
-                    let result_path = local_queue::result_path(&group_dir, request.job_id);
-                    std::fs::create_dir_all(result_path.parent().unwrap()).unwrap();
-                    std::fs::write(&result_path, serde_json::to_vec(&response).unwrap()).unwrap();
-                    return request;
-                }
-            }
-
-            tokio::task::yield_now().await;
-        }
-    }
-
-    async fn wait_for_job_and_write_success(
-        group_dir: std::path::PathBuf,
-        profile: String,
-    ) -> JobRequest {
-        wait_for_job_and_write_result(group_dir, profile, 0, None).await
-    }
-
-    async fn wait_for_active_inputs_and_write_success(
-        group_dir: std::path::PathBuf,
-        profile: String,
-        expected_inputs: usize,
-    ) -> (JobRequest, Vec<local_queue::ActiveInputEntry>) {
-        let job_dir = local_queue::profile_jobs_dir(&group_dir, &profile).unwrap();
-        let queue = local_queue::LocalQueue::new(group_dir.clone());
-        let deadline = tokio::time::Instant::now() + TEST_QUEUE_WATCH_TIMEOUT;
-        let mut last_seen_inputs = 0;
-        loop {
-            if let Ok(entries) = std::fs::read_dir(&job_dir) {
-                for entry in entries.filter_map(Result::ok) {
-                    let path = entry.path();
-                    if path.extension().and_then(|ext| ext.to_str()) != Some("job") {
-                        continue;
-                    }
-                    let request: JobRequest =
-                        serde_json::from_slice(&std::fs::read(&path).unwrap()).unwrap();
-                    let inputs =
-                        queue.read_active_input_entries_from_sequence_sync(request.job_id, 0);
-                    last_seen_inputs = last_seen_inputs.max(inputs.len());
-                    if inputs.len() < expected_inputs {
-                        continue;
-                    }
-                    let response = JobResponse {
-                        run_id: request.job_id,
-                        exit_code: 0,
-                        error: None,
-                    };
-                    let result_path = local_queue::result_path(&group_dir, request.job_id);
-                    std::fs::create_dir_all(result_path.parent().unwrap()).unwrap();
-                    std::fs::write(&result_path, serde_json::to_vec(&response).unwrap()).unwrap();
-                    return (request, inputs);
-                }
-            }
-
-            if tokio::time::Instant::now() >= deadline {
-                panic!(
-                    "timed out waiting for {expected_inputs} local active inputs in {} (last seen: {last_seen_inputs})",
-                    job_dir.display()
-                );
-            }
-            tokio::time::sleep(TEST_QUEUE_WATCH_INTERVAL).await;
-        }
-    }
-
-    fn submit_args_for_test() -> SubmitArgs {
-        SubmitArgs {
-            group: "test/group".into(),
-            prompt: "hello".into(),
-            cli_agent_type: "claude-code".into(),
-            profile: None,
-            session_id: None,
-            feature_flags: vec![],
-            env: vec![],
-            secret_env: vec![],
-            timeout: 1,
-            active_inputs: vec![],
-        }
-    }
-
-    #[tokio::test]
-    async fn submit_defaults_profile_and_writes_default_partition() {
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().to_path_buf());
-        let group = "test/group";
-        let group_dir = home.groups_dir().join(group);
-        let watcher = tokio::spawn(wait_for_job_and_write_success(
-            group_dir,
-            crate::profile::DEFAULT_PROFILE.to_owned(),
-        ));
-
-        let code = run_submit_with_home(
-            SubmitArgs {
-                group: group.into(),
-                prompt: "hello".into(),
-                cli_agent_type: "claude-code".into(),
-                profile: None,
-                session_id: None,
-                feature_flags: vec![],
-                env: vec![],
-                secret_env: vec![],
-                timeout: 5,
-                active_inputs: vec![],
-            },
-            home,
-        )
-        .await
-        .unwrap();
-        let request = watcher.await.unwrap();
-
-        assert_eq!(code, ExitCode::SUCCESS);
-        assert_eq!(request.prompt, "hello");
-        assert_eq!(request.cli_agent_type, "claude-code");
-        assert_eq!(
-            request.profile.as_deref(),
-            Some(crate::profile::DEFAULT_PROFILE)
-        );
-    }
-
-    #[tokio::test]
-    async fn submit_writes_non_default_profile_partition() {
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().to_path_buf());
-        let group = "test/group";
-        let profile = "vm0/large";
-        let group_dir = home.groups_dir().join(group);
-        let watcher = tokio::spawn(wait_for_job_and_write_success(
-            group_dir,
-            profile.to_owned(),
-        ));
-
-        let code = run_submit_with_home(
-            SubmitArgs {
-                group: group.into(),
-                prompt: "hello".into(),
-                cli_agent_type: "claude-code".into(),
-                profile: Some(profile.into()),
-                session_id: None,
-                feature_flags: vec![],
-                env: vec![],
-                secret_env: vec![],
-                timeout: 5,
-                active_inputs: vec![],
-            },
-            home,
-        )
-        .await
-        .unwrap();
-        let request = watcher.await.unwrap();
-
-        assert_eq!(code, ExitCode::SUCCESS);
-        assert_eq!(request.profile.as_deref(), Some(profile));
-    }
-
-    #[tokio::test]
-    async fn submit_serializes_feature_flags() {
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().to_path_buf());
-        let group = "test/group";
-        let group_dir = home.groups_dir().join(group);
-        let watcher = tokio::spawn(wait_for_job_and_write_success(
-            group_dir,
-            crate::profile::DEFAULT_PROFILE.to_owned(),
-        ));
-
-        let code = run_submit_with_home(
-            SubmitArgs {
-                group: group.into(),
-                prompt: "hello".into(),
-                cli_agent_type: "codex".into(),
-                profile: None,
-                session_id: Some("sess-123".into()),
-                feature_flags: vec!["alpha=true".into(), "beta=false".into()],
-                env: vec![],
-                secret_env: vec![],
-                timeout: 5,
-                active_inputs: vec![],
-            },
-            home,
-        )
-        .await
-        .unwrap();
-        let request = watcher.await.unwrap();
-        let flags = request.feature_flags.as_ref().unwrap();
-
-        assert_eq!(code, ExitCode::SUCCESS);
-        assert_eq!(request.prompt, "hello");
-        assert_eq!(request.cli_agent_type, "codex");
-        assert_eq!(request.session_id.as_deref(), Some("sess-123"));
-        assert_eq!(flags.get("alpha"), Some(&true));
-        assert_eq!(flags.get("beta"), Some(&false));
-    }
-
-    #[test]
-    fn parses_active_input_specs() {
-        let job_id = RunId::nil();
-        let parsed = SubmitPlan::parse_active_inputs(
-            &[
-                "after=250ms,text=first".to_string(),
-                "after=1s,text=second,with,commas".to_string(),
-            ],
-            Duration::from_secs(5),
-            job_id,
-        )
-        .unwrap();
-
-        assert_eq!(
-            parsed,
-            vec![
-                DelayedActiveInput {
-                    sequence: 1,
-                    message_id: format!("local-active-input-{job_id}-1"),
-                    after: Duration::from_millis(250),
-                    text: "first".to_string(),
-                },
-                DelayedActiveInput {
-                    sequence: 2,
-                    message_id: format!("local-active-input-{job_id}-2"),
-                    after: Duration::from_secs(1),
-                    text: "second,with,commas".to_string(),
-                },
-            ]
-        );
-    }
-
-    #[test]
-    fn parses_more_than_eight_active_input_specs() {
-        let job_id = RunId::nil();
-        let values = (1..=9)
-            .map(|sequence| format!("after={sequence}ms,text=input-{sequence}"))
-            .collect::<Vec<_>>();
-        let parsed =
-            SubmitPlan::parse_active_inputs(&values, Duration::from_secs(5), job_id).unwrap();
-
-        assert_eq!(parsed.len(), 9);
-        assert_eq!(parsed[8].sequence, 9);
-        assert_eq!(
-            parsed[8].message_id,
-            format!("local-active-input-{job_id}-9")
-        );
-        assert_eq!(parsed[8].text, "input-9");
-    }
-
-    #[test]
-    fn rejects_invalid_active_input_specs() {
-        let job_id = RunId::nil();
-        for value in [
-            "text=missing-after",
-            "after=1m,text=bad-unit",
-            "after=0s,text=zero",
-            "after=1s,text=",
-            "after=5s,text=timeout",
-            "after=1s,text=bad\0nul",
-        ] {
-            let err = SubmitPlan::parse_active_inputs(
-                &[value.to_string()],
-                Duration::from_secs(5),
-                job_id,
-            )
-            .unwrap_err();
-
-            assert!(
-                err.to_string().contains("active-input"),
-                "value={value}, got: {err}"
-            );
-        }
-
-        let err = SubmitPlan::parse_active_inputs(
-            &[
-                "after=2s,text=first".to_string(),
-                "after=1s,text=second".to_string(),
-            ],
-            Duration::from_secs(5),
-            job_id,
-        )
-        .unwrap_err();
-        assert!(err.to_string().contains("non-decreasing"));
-
-        let oversized_text = "x".repeat(ACTIVE_INPUT_CONTROL_PAYLOAD_MAX_BYTES);
-        let err = SubmitPlan::parse_active_inputs(
-            &[format!("after=1s,text={oversized_text}")],
-            Duration::from_secs(5),
-            job_id,
-        )
-        .unwrap_err();
-        assert!(err.to_string().contains("serialized payload"));
-    }
-
-    #[test]
-    fn accepts_active_input_payload_at_serialized_limit() {
-        let job_id = RunId::nil();
-        let payload_overhead = active_input_payload_len("").unwrap();
-        let exact_limit_text =
-            "x".repeat(ACTIVE_INPUT_CONTROL_PAYLOAD_MAX_BYTES - payload_overhead);
-
-        let parsed = SubmitPlan::parse_active_inputs(
-            &[format!("after=1s,text={exact_limit_text}")],
-            Duration::from_secs(5),
-            job_id,
-        )
-        .unwrap();
-
-        assert_eq!(parsed.len(), 1);
-        assert_eq!(
-            active_input_payload_len(&parsed[0].text).unwrap(),
-            ACTIVE_INPUT_CONTROL_PAYLOAD_MAX_BYTES
-        );
-    }
-
-    #[test]
-    fn accepts_active_input_for_codex_agent() {
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().to_path_buf());
-        let mut args = submit_args_for_test();
-        args.cli_agent_type = "codex".to_string();
-        args.active_inputs = vec!["after=1ms,text=hello".to_string()];
-
-        let plan = SubmitPlan::from_args(args, home).unwrap();
-        let request: JobRequest = serde_json::from_slice(&plan.request_json).unwrap();
-        assert_eq!(request.cli_agent_type, "codex");
-        assert_eq!(request.active_input, Some(true));
-        assert_eq!(plan.active_inputs.len(), 1);
-    }
-
-    #[test]
-    fn accepts_active_input_for_custom_agent_that_defaults_to_claude() {
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().to_path_buf());
-        let mut args = submit_args_for_test();
-        args.cli_agent_type = "custom-agent".to_string();
-        args.active_inputs = vec!["after=1ms,text=hello".to_string()];
-
-        SubmitPlan::from_args(args, home).unwrap();
-    }
-
-    #[tokio::test]
-    async fn active_inputs_are_written_after_job_publication_and_cleaned_on_completion() {
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().to_path_buf());
-        let group = "test/group";
-        let group_dir = home.groups_dir().join(group);
-        let watcher = tokio::spawn(wait_for_active_inputs_and_write_success(
-            group_dir.clone(),
-            crate::profile::DEFAULT_PROFILE.to_owned(),
-            9,
-        ));
-        let mut args = submit_args_for_test();
-        args.group = group.into();
-        args.timeout = 5;
-        args.active_inputs = (1..=9)
-            .map(|sequence| {
-                let text = if sequence == 2 {
-                    "second,with,comma".to_string()
-                } else {
-                    format!("input-{sequence}")
-                };
-                format!("after={sequence}ms,text={text}")
-            })
-            .collect();
-        let expected_inputs = (1..=9)
-            .map(|sequence| {
-                let text = if sequence == 2 {
-                    "second,with,comma".to_string()
-                } else {
-                    format!("input-{sequence}")
-                };
-                (sequence, text)
-            })
-            .collect::<Vec<_>>();
-
-        let code = run_submit_with_home(args, home).await.unwrap();
-        let (request, inputs) = watcher.await.unwrap();
-
-        assert_eq!(code, ExitCode::SUCCESS);
-        assert_eq!(request.prompt, "hello");
-        assert_eq!(request.active_input, Some(true));
-        assert_eq!(
-            inputs
-                .iter()
-                .map(|entry| (entry.sequence, entry.text.clone()))
-                .collect::<Vec<_>>(),
-            expected_inputs
-        );
-        assert!(
-            !local_queue::run_inputs_dir(&group_dir, request.job_id).exists(),
-            "completed submit cleanup should remove local active-input files"
-        );
-    }
-
-    #[tokio::test]
-    async fn active_input_producer_stops_after_write_failure_to_preserve_sequence_order() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        write_queue_job_file(group_dir, crate::profile::DEFAULT_PROFILE, job_id);
-        local_queue::ensure_run_inputs_dir(group_dir, job_id).unwrap();
-        std::fs::write(
-            local_queue::active_input_path(group_dir, job_id, 1),
-            b"not-json",
-        )
-        .unwrap();
-
-        let producer = ActiveInputProducer::start(
-            queue,
-            vec![
-                DelayedActiveInput {
-                    sequence: 1,
-                    message_id: "msg-1".to_string(),
-                    after: Duration::ZERO,
-                    text: "first".to_string(),
-                },
-                DelayedActiveInput {
-                    sequence: 2,
-                    message_id: "msg-2".to_string(),
-                    after: Duration::ZERO,
-                    text: "second".to_string(),
-                },
-            ],
-        )
-        .unwrap();
-
-        tokio::time::timeout(TEST_QUEUE_WATCH_TIMEOUT, producer.task)
-            .await
-            .unwrap()
-            .unwrap();
-
-        assert!(
-            !local_queue::active_input_path(group_dir, job_id, 2).exists(),
-            "producer should not create a sequence gap after an active-input write failure"
-        );
-    }
-
-    #[tokio::test]
-    async fn submit_serializes_env_and_secret_env() {
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().to_path_buf());
-        let group = "test/group";
-        let group_dir = home.groups_dir().join(group);
-        let watcher = tokio::spawn(wait_for_job_and_write_success(
-            group_dir,
-            crate::profile::DEFAULT_PROFILE.to_owned(),
-        ));
-
-        let mut args = submit_args_for_test();
-        args.group = group.into();
-        args.timeout = 5;
-        args.env = vec![
-            "FOO=bar".into(),
-            "URL=https://example.test/path?a=1&b=2".into(),
-            "EMPTY=".into(),
-            "MULTILINE=line1\nline2".into(),
-            "VM0_STUCK_TOOL_TIMEOUT_SECS=3".into(),
-        ];
-        args.secret_env = vec![
-            "ANTHROPIC_API_KEY=sk-ant-local-secret".into(),
-            "PRIVATE_KEY=-----BEGIN KEY-----\r\nsecret\r\n-----END KEY-----".into(),
-        ];
-
-        let code = run_submit_with_home(args, home).await.unwrap();
-        let request = watcher.await.unwrap();
-        let environment = request.environment.as_ref().unwrap();
-        let secret_environment = request.secret_environment.as_ref().unwrap();
-
-        assert_eq!(code, ExitCode::SUCCESS);
-        assert_eq!(environment.get("FOO").map(String::as_str), Some("bar"));
-        assert_eq!(
-            environment.get("URL").map(String::as_str),
-            Some("https://example.test/path?a=1&b=2")
-        );
-        assert_eq!(environment.get("EMPTY").map(String::as_str), Some(""));
-        assert_eq!(
-            environment.get("MULTILINE").map(String::as_str),
-            Some("line1\nline2")
-        );
-        assert_eq!(
-            environment
-                .get("VM0_STUCK_TOOL_TIMEOUT_SECS")
-                .map(String::as_str),
-            Some("3")
-        );
-        assert_eq!(
-            secret_environment
-                .get("ANTHROPIC_API_KEY")
-                .map(String::as_str),
-            Some("sk-ant-local-secret")
-        );
-        assert_eq!(
-            secret_environment.get("PRIVATE_KEY").map(String::as_str),
-            Some("-----BEGIN KEY-----\r\nsecret\r\n-----END KEY-----")
-        );
-    }
-
-    #[tokio::test]
-    async fn rejects_invalid_env_entries_before_submit() {
-        let cases = vec![
-            (vec!["FOO".to_string()], Vec::new(), "expected KEY=VALUE"),
-            (vec!["=VALUE".to_string()], Vec::new(), "expected KEY=VALUE"),
-            (
-                vec!["BAD-KEY=value".to_string()],
-                Vec::new(),
-                "expected [_A-Za-z][_A-Za-z0-9]*",
-            ),
-            (
-                vec!["1KEY=value".to_string()],
-                Vec::new(),
-                "expected [_A-Za-z][_A-Za-z0-9]*",
-            ),
-            (
-                vec!["KEY SPACE=value".to_string()],
-                Vec::new(),
-                "expected [_A-Za-z][_A-Za-z0-9]*",
-            ),
-            (
-                Vec::new(),
-                vec!["ÅKEY=value".to_string()],
-                "expected [_A-Za-z][_A-Za-z0-9]*",
-            ),
-            (
-                Vec::new(),
-                vec!["KEY=with\0nul".to_string()],
-                "NUL characters",
-            ),
-            (
-                vec!["VM0_PROMPT=value".to_string()],
-                Vec::new(),
-                "runner-owned environment variables",
-            ),
-            (
-                Vec::new(),
-                vec!["CLI_AGENT_TYPE=codex".to_string()],
-                "runner-owned environment variables",
-            ),
-            (
-                Vec::new(),
-                vec!["VM0_STUCK_TOOL_TIMEOUT_SECS=3".to_string()],
-                "must be passed with --env",
-            ),
-            (
-                vec!["FOO=1".to_string(), "FOO=2".to_string()],
-                Vec::new(),
-                "duplicate --env key 'FOO'",
-            ),
-            (
-                vec!["FOO=1".to_string()],
-                vec!["FOO=2".to_string()],
-                "across --env and --secret-env",
-            ),
-        ];
-
-        for (env, secret_env, expected) in cases {
-            let dir = tempfile::tempdir().unwrap();
-            let home = HomePaths::with_root(dir.path().to_path_buf());
-            let mut args = submit_args_for_test();
-            args.env = env;
-            args.secret_env = secret_env;
-
-            let err = run_submit_with_home(args, home).await.unwrap_err();
-
-            assert!(err.to_string().contains(expected), "got: {err}");
-        }
-    }
-
-    #[test]
-    fn write_job_file_creates_private_job_file() {
-        use std::os::unix::fs::PermissionsExt;
-
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.job.parent().unwrap()).unwrap();
-        let plan = SubmitPlan {
-            group: "test/group".into(),
-            profile: crate::profile::DEFAULT_PROFILE.to_owned(),
-            queue,
-            timeout: Duration::ZERO,
-            request_json: br#"{"secretEnvironment":{"ANTHROPIC_API_KEY":"sk-local-secret"}}"#
-                .to_vec(),
-            active_inputs: vec![],
-        };
-
-        plan.write_job_file().unwrap();
-
-        let mode = std::fs::metadata(&plan.queue.job)
-            .unwrap()
-            .permissions()
-            .mode()
-            & 0o777;
-        assert_eq!(mode, crate::host_file::PRIVATE_FILE_MODE);
-    }
-
-    #[test]
-    fn write_job_file_removes_tmp_when_publish_fails() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.job.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(&queue.job).unwrap();
-        let plan = SubmitPlan {
-            group: "test/group".into(),
-            profile: crate::profile::DEFAULT_PROFILE.to_owned(),
-            queue,
-            timeout: Duration::ZERO,
-            request_json: b"{}".to_vec(),
-            active_inputs: vec![],
-        };
-
-        let err = plan.write_job_file().unwrap_err();
-
-        assert!(err.to_string().contains("rename job file"), "got: {err}");
-        assert!(plan.queue.job.is_dir());
-        let tmp_files: Vec<_> = std::fs::read_dir(&plan.queue.job_dir)
-            .unwrap()
-            .filter_map(Result::ok)
-            .filter(|entry| entry.path().extension().and_then(|ext| ext.to_str()) == Some("tmp"))
-            .collect();
-        assert!(tmp_files.is_empty(), "tmp files left behind: {tmp_files:?}");
-    }
-
-    #[test]
-    fn submit_plan_creates_private_queue_dirs_and_job_file() {
-        let dir = tempfile::tempdir().unwrap();
-        let root = dir.path().to_path_buf();
-        let home = HomePaths::with_root(root.clone());
-        let group = "test/group";
-        let group_dir = root.join("groups").join(group);
-        let plan = SubmitPlan::from_args(
-            SubmitArgs {
-                group: group.into(),
-                prompt: "secret prompt".into(),
-                cli_agent_type: "claude-code".into(),
-                profile: None,
-                session_id: Some("session-123".into()),
-                feature_flags: vec![],
-                env: vec![],
-                secret_env: vec![],
-                timeout: 5,
-                active_inputs: vec![],
-            },
-            home,
-        )
-        .unwrap();
-
-        plan.write_job_file().unwrap();
-
-        assert_eq!(mode(&plan.queue.job_dir), 0o700);
-        assert_eq!(mode(&local_queue::results_dir(&group_dir)), 0o700);
-        assert_eq!(mode(&local_queue::cancels_dir(&group_dir)), 0o700);
-        assert_eq!(mode(&plan.queue.job), 0o600);
-    }
-
-    #[test]
-    fn submit_plan_tightens_existing_permissive_queue_dirs() {
-        let dir = tempfile::tempdir().unwrap();
-        let root = dir.path().to_path_buf();
-        let home = HomePaths::with_root(root.clone());
-        let group = "test/group";
-        let group_dir = root.join("groups").join(group);
-        let job_dir =
-            local_queue::profile_jobs_dir(&group_dir, crate::profile::DEFAULT_PROFILE).unwrap();
-        let results_dir = local_queue::results_dir(&group_dir);
-        let cancels_dir = local_queue::cancels_dir(&group_dir);
-        for path in [&job_dir, &results_dir, &cancels_dir] {
-            std::fs::create_dir_all(path).unwrap();
-            std::fs::set_permissions(path, std::fs::Permissions::from_mode(0o755)).unwrap();
-        }
-
-        let plan = SubmitPlan::from_args(
-            SubmitArgs {
-                group: group.into(),
-                prompt: "secret prompt".into(),
-                cli_agent_type: "claude-code".into(),
-                profile: None,
-                session_id: None,
-                feature_flags: vec![],
-                env: vec![],
-                secret_env: vec![],
-                timeout: 5,
-                active_inputs: vec![],
-            },
-            home,
-        )
-        .unwrap();
-
-        assert_eq!(mode(&plan.queue.job_dir), 0o700);
-        assert_eq!(mode(&results_dir), 0o700);
-        assert_eq!(mode(&cancels_dir), 0o700);
-    }
-
-    #[tokio::test]
-    async fn submit_returns_failure_for_nonzero_job_response() {
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().to_path_buf());
-        let group = "test/group";
-        let group_dir = home.groups_dir().join(group);
-        let watcher = tokio::spawn(wait_for_job_and_write_result(
-            group_dir.clone(),
-            crate::profile::DEFAULT_PROFILE.to_owned(),
-            42,
-            Some("agent failed".into()),
-        ));
-
-        let code = run_submit_with_home(
-            SubmitArgs {
-                group: group.into(),
-                prompt: "hello".into(),
-                cli_agent_type: "claude-code".into(),
-                profile: None,
-                session_id: None,
-                feature_flags: vec![],
-                env: vec![],
-                secret_env: vec![],
-                timeout: 5,
-                active_inputs: vec![],
-            },
-            home,
-        )
-        .await
-        .unwrap();
-        let request = watcher.await.unwrap();
-        let result_path = local_queue::result_path(&group_dir, request.job_id);
-
-        assert_eq!(code, ExitCode::FAILURE);
-        assert!(
-            !result_path.exists(),
-            "completed cleanup should remove nonzero result files"
-        );
-    }
-
-    #[test]
-    fn cleanup_completed_is_idempotent() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.job.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.result.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.claim.parent().unwrap()).unwrap();
-
-        // Create some files
-        std::fs::write(&queue.job, b"{}").unwrap();
-        std::fs::write(&queue.result, b"{}").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-        std::fs::write(&queue.claim, b"").unwrap();
-
-        // First cleanup
-        queue.cleanup_completed();
-        assert!(!queue.job.exists());
-        assert!(!queue.result.exists());
-        assert!(!queue.cancel.exists());
-        assert!(
-            !queue.claim.exists(),
-            "completed-result cleanup should remove stale claims left after result write"
-        );
-
-        // Second cleanup (idempotent — no panic on missing files)
-        queue.cleanup_completed();
-    }
-
-    #[test]
-    fn completed_cleanup_removes_duplicate_job_files() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        let default_job = write_queue_job_file(group_dir, crate::profile::DEFAULT_PROFILE, job_id);
-        let large_job = write_queue_job_file(group_dir, "vm0/large", job_id);
-        std::fs::create_dir_all(queue.result.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.claim.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.result, b"{}").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-        std::fs::write(&queue.claim, b"").unwrap();
-
-        queue.cleanup_completed();
-
-        assert!(!default_job.exists());
-        assert!(!large_job.exists());
-        assert!(!queue.result.exists());
-        assert!(!queue.cancel.exists());
-        assert!(!queue.claim.exists());
-    }
-
-    #[test]
-    fn completed_cleanup_keeps_result_when_job_cannot_be_removed() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(&queue.job).unwrap();
-        std::fs::create_dir_all(queue.result.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.claim.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.result, b"{}").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-        std::fs::write(&queue.claim, b"").unwrap();
-
-        queue.cleanup_completed();
-
-        assert!(
-            queue.result.exists(),
-            "result must remain as the terminal marker if the job path was not removed"
-        );
-        assert!(!queue.cancel.exists());
-        assert!(!queue.claim.exists());
-    }
-
-    #[test]
-    fn completed_cleanup_keeps_result_when_duplicate_job_cannot_be_removed() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        let default_job = write_queue_job_file(group_dir, crate::profile::DEFAULT_PROFILE, job_id);
-        let blocked_job = local_queue::job_path(group_dir, "vm0/large", job_id).unwrap();
-        std::fs::create_dir_all(&blocked_job).unwrap();
-        std::fs::create_dir_all(queue.result.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.claim.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.result, b"{}").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-        std::fs::write(&queue.claim, b"").unwrap();
-
-        queue.cleanup_completed();
-
-        assert!(!default_job.exists());
-        assert!(blocked_job.exists());
-        assert!(
-            queue.result.exists(),
-            "result must remain as the terminal marker if any duplicate job path was not removed"
-        );
-        assert!(!queue.cancel.exists());
-        assert!(!queue.claim.exists());
-    }
-
-    #[test]
-    fn completed_cleanup_removes_result_when_job_already_absent() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.result.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.claim.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.result, b"{}").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-        std::fs::write(&queue.claim, b"").unwrap();
-
-        queue.cleanup_completed();
-
-        assert!(!queue.result.exists());
-        assert!(!queue.cancel.exists());
-        assert!(!queue.claim.exists());
-    }
-
-    #[test]
-    fn abandoned_cleanup_preserves_active_claim_state() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.job.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.result.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.claim.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.job, b"{}").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-        std::fs::write(&queue.claim, b"").unwrap();
-        local_queue::LocalQueue::new(group_dir.to_path_buf())
-            .write_active_input_sync(&local_queue::ActiveInputEntry {
-                run_id: job_id,
-                sequence: 1,
-                message_id: "msg-1".to_string(),
-                text: "one".to_string(),
-            })
-            .unwrap();
-
-        queue.abandon("timed out");
-        let response: JobResponse =
-            serde_json::from_slice(&std::fs::read(&queue.result).unwrap()).unwrap();
-
-        assert_eq!(response.run_id, queue.job_id);
-        assert!(!queue.job.exists());
-        assert!(
-            queue.result.exists(),
-            "abandoned cleanup must keep a terminal marker while a runner owns the claim"
-        );
-        assert!(
-            queue.cancel.exists(),
-            "abandoned cleanup must not delete files while a runner owns the claim"
-        );
-        assert!(queue.claim.exists());
-        assert!(
-            local_queue::run_inputs_dir(group_dir, job_id).exists(),
-            "abandoned cleanup must leave active inputs for claimed jobs"
-        );
-    }
-
-    #[test]
-    fn abandoned_cleanup_removes_unclaimed_job_without_claim_marker() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.job.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.result.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.job, b"{}").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-        local_queue::LocalQueue::new(group_dir.to_path_buf())
-            .write_active_input_sync(&local_queue::ActiveInputEntry {
-                run_id: job_id,
-                sequence: 1,
-                message_id: "msg-1".to_string(),
-                text: "one".to_string(),
-            })
-            .unwrap();
-
-        queue.abandon("timed out");
-
-        assert!(!queue.job.exists());
-        assert!(!queue.result.exists());
-        assert!(!queue.cancel.exists());
-        assert!(
-            !queue.claim.exists(),
-            "abandoned cleanup should not create a temporary claim"
-        );
-        assert!(
-            !local_queue::run_inputs_dir(group_dir, job_id).exists(),
-            "abandoned cleanup should remove active inputs for unclaimed jobs"
-        );
-    }
-
-    #[test]
-    fn abandoned_cleanup_ignores_claim_file_symlink() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.job.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.claim.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.job, b"{}").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-        let target = dir.path().join("target-claim");
-        std::fs::write(&target, b"").unwrap();
-        symlink(&target, &queue.claim).unwrap();
-
-        queue.abandon("timed out");
-
-        assert!(!queue.job.exists());
-        assert!(!queue.result.exists());
-        assert!(!queue.cancel.exists());
-        assert!(!queue.claim.exists());
-    }
-
-    #[test]
-    fn abandoned_cleanup_removes_duplicate_unclaimed_jobs() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        let default_job = write_queue_job_file(group_dir, crate::profile::DEFAULT_PROFILE, job_id);
-        let large_job = write_queue_job_file(group_dir, "vm0/large", job_id);
-        std::fs::create_dir_all(queue.result.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.cancel, b"").unwrap();
-
-        queue.abandon("timed out");
-
-        assert!(!default_job.exists());
-        assert!(!large_job.exists());
-        assert!(!queue.result.exists());
-        assert!(!queue.cancel.exists());
-        assert!(
-            !queue.claim.exists(),
-            "abandoned cleanup should not create a temporary claim"
-        );
-    }
-
-    #[test]
-    fn abandoned_cleanup_removes_marker_when_job_already_absent() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.cancel, b"").unwrap();
-        let marker =
-            write_abandoned_result_marker(&queue.result, job_id, "local submit abandoned").unwrap();
-
-        queue.cleanup_abandoned(Some(&marker));
-
-        assert!(!queue.result.exists());
-        assert!(!queue.cancel.exists());
-        assert!(!queue.claim.exists());
-    }
-
-    #[test]
-    fn abandoned_cleanup_removes_unclaimed_active_inputs_when_job_already_absent() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-        let marker =
-            write_abandoned_result_marker(&queue.result, job_id, "local submit abandoned").unwrap();
-        local_queue::LocalQueue::new(group_dir.to_path_buf())
-            .write_active_input_sync(&local_queue::ActiveInputEntry {
-                run_id: job_id,
-                sequence: 1,
-                message_id: "msg-1".to_string(),
-                text: "one".to_string(),
-            })
-            .unwrap();
-
-        queue.cleanup_abandoned(Some(&marker));
-
-        assert!(!local_queue::run_inputs_dir(group_dir, job_id).exists());
-        assert!(!queue.result.exists());
-        assert!(!queue.cancel.exists());
-        assert!(!queue.claim.exists());
-    }
-
-    #[test]
-    fn abandoned_cleanup_removes_late_unclaimed_active_inputs_after_job_cleanup() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.job.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-        std::fs::write(&queue.job, b"{}").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-
-        queue.abandon("timed out");
-        local_queue::LocalQueue::new(group_dir.to_path_buf())
-            .write_active_input_sync(&local_queue::ActiveInputEntry {
-                run_id: job_id,
-                sequence: 1,
-                message_id: "msg-1".to_string(),
-                text: "late".to_string(),
-            })
-            .unwrap();
-
-        queue.cleanup_abandoned(None);
-
-        assert!(!local_queue::run_inputs_dir(group_dir, job_id).exists());
-        assert!(!queue.result.exists());
-        assert!(!queue.cancel.exists());
-        assert!(!queue.claim.exists());
-    }
-
-    #[test]
-    fn abandoned_cleanup_keeps_marker_when_duplicate_job_cannot_be_removed() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        let default_job = write_queue_job_file(group_dir, crate::profile::DEFAULT_PROFILE, job_id);
-        let blocked_job = local_queue::job_path(group_dir, "vm0/large", job_id).unwrap();
-        std::fs::create_dir_all(&blocked_job).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.cancel, b"").unwrap();
-
-        queue.abandon("timed out");
-
-        assert!(!default_job.exists());
-        assert!(blocked_job.exists());
-        assert!(
-            queue.result.exists(),
-            "terminal marker must remain if any duplicate job path could not be removed"
-        );
-        assert!(queue.cancel.exists());
-        assert!(!queue.claim.exists());
-    }
-
-    #[test]
-    fn abandoned_cleanup_keeps_marker_when_job_already_absent_but_claimed() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.claim.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.cancel, b"").unwrap();
-        std::fs::write(&queue.claim, b"").unwrap();
-
-        queue.abandon("timed out");
-
-        assert!(queue.result.exists());
-        assert!(queue.cancel.exists());
-        assert!(queue.claim.exists());
-    }
-
-    #[test]
-    fn abandoned_cleanup_removes_stale_empty_result_after_unclaimed_job() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.job.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.result.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.job, b"{}").unwrap();
-        std::fs::write(&queue.result, b"").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-
-        queue.abandon("timed out");
-
-        assert!(!queue.job.exists());
-        assert!(
-            !queue.result.exists(),
-            "empty stale result should not strand an unclaimed abandoned job"
-        );
-        assert!(!queue.cancel.exists());
-        assert!(!queue.claim.exists());
-    }
-
-    #[test]
-    fn abandoned_cleanup_keeps_runner_result_published_over_empty_result() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.job.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.result.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.job, b"{}").unwrap();
-        std::fs::write(&queue.result, b"").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-        let marker = write_abandoned_result_marker(&queue.result, job_id, "local submit abandoned");
-        assert!(marker.is_none());
-
-        let runner_queue = local_queue::LocalQueue::new(group_dir.to_path_buf());
-        assert!(runner_queue.write_result_sync(job_id, 0, None));
-
-        queue.cleanup_abandoned(None);
-
-        assert!(!queue.job.exists());
-        let response: JobResponse =
-            serde_json::from_slice(&std::fs::read(&queue.result).unwrap()).unwrap();
-        assert_eq!(response.run_id, job_id);
-        assert_eq!(response.exit_code, 0);
-        assert!(response.error.is_none());
-        assert!(!queue.cancel.exists());
-        assert!(!queue.claim.exists());
-    }
-
-    #[test]
-    fn abandoned_cleanup_removes_unclaimed_job_when_marker_cannot_be_written() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        let result_dir = local_queue::results_dir(group_dir);
-        std::fs::create_dir_all(queue.job.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.job, b"{}").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-        std::fs::write(&result_dir, b"not a directory").unwrap();
-
-        queue.abandon("timed out");
-
-        assert!(
-            !queue.job.exists(),
-            "timed-out unclaimed job should not remain executable after marker write failure"
-        );
-        assert!(!queue.cancel.exists());
-        assert!(!queue.claim.exists());
-        assert!(result_dir.is_file());
-    }
-
-    #[test]
-    fn abandoned_cleanup_keeps_completed_result() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.job.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.result.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.job, b"{}").unwrap();
-        std::fs::write(&queue.result, b"{}").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-
-        queue.abandon("timed out");
-
-        assert!(!queue.job.exists());
-        assert!(
-            queue.result.exists(),
-            "abandoned cleanup must not delete a non-empty result written by a runner"
-        );
-        assert!(!queue.cancel.exists());
-        assert!(!queue.claim.exists());
-    }
-
-    #[test]
-    fn abandoned_cleanup_keeps_completed_result_when_claimed() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(queue.job.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.result.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(queue.claim.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.job, b"{}").unwrap();
-        std::fs::write(&queue.result, b"{}").unwrap();
-        std::fs::write(&queue.cancel, b"").unwrap();
-        std::fs::write(&queue.claim, b"").unwrap();
-
-        queue.abandon("timed out");
-
-        assert!(!queue.job.exists());
-        assert!(queue.result.exists());
-        assert!(queue.cancel.exists());
-        assert!(queue.claim.exists());
-    }
-
-    #[test]
-    fn abandoned_cleanup_keeps_marker_when_job_cannot_be_removed() {
-        let dir = tempfile::tempdir().unwrap();
-        let group_dir = dir.path();
-        let job_id = RunId::new_v4();
-        let queue = submit_queue_entry(group_dir, job_id);
-        std::fs::create_dir_all(&queue.job).unwrap();
-        std::fs::create_dir_all(queue.cancel.parent().unwrap()).unwrap();
-
-        std::fs::write(&queue.cancel, b"").unwrap();
-
-        queue.abandon("timed out");
-
-        assert!(
-            queue.result.exists(),
-            "terminal marker must remain if the stale job path could not be removed"
-        );
-        assert!(queue.cancel.exists());
-        assert!(!queue.claim.exists());
-    }
-
-    #[tokio::test]
-    async fn rejects_invalid_profile_name() {
-        let args = SubmitArgs {
-            group: "test/group".into(),
-            prompt: "hello".into(),
-            cli_agent_type: "claude-code".into(),
-            profile: Some("bad-name".into()),
-            session_id: None,
-            feature_flags: vec![],
-            env: vec![],
-            secret_env: vec![],
-            timeout: 1,
-            active_inputs: vec![],
-        };
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().to_path_buf());
-        let err = run_submit_with_home(args, home).await.unwrap_err();
-        assert!(
-            err.to_string().contains("invalid profile name"),
-            "got: {err}"
-        );
-    }
-
-    #[tokio::test]
-    async fn accepts_valid_profile_name() {
-        let args = SubmitArgs {
-            group: "test/group".into(),
-            prompt: "hello".into(),
-            cli_agent_type: "claude-code".into(),
-            profile: Some("vm0/default".into()),
-            session_id: None,
-            feature_flags: vec![],
-            env: vec![],
-            secret_env: vec![],
-            timeout: 0,
-            active_inputs: vec![],
-        };
-        // Should pass validation and fail later (HomePaths or timeout), not on profile.
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().to_path_buf());
-        let result = run_submit_with_home(args, home).await;
-        if let Err(e) = &result {
-            assert!(!e.to_string().contains("invalid profile name"), "got: {e}");
-        }
-    }
-
-    #[tokio::test]
-    async fn rejects_feature_flag_missing_equals() {
-        let args = SubmitArgs {
-            group: "test/group".into(),
-            prompt: "hello".into(),
-            cli_agent_type: "claude-code".into(),
-            profile: None,
-            session_id: None,
-            feature_flags: vec!["myFlag".into()],
-            env: vec![],
-            secret_env: vec![],
-            timeout: 1,
-            active_inputs: vec![],
-        };
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().to_path_buf());
-        let err = run_submit_with_home(args, home).await.unwrap_err();
-        assert!(err.to_string().contains("expected key=value"), "got: {err}");
-    }
-
-    #[tokio::test]
-    async fn rejects_feature_flag_non_boolean() {
-        let args = SubmitArgs {
-            group: "test/group".into(),
-            prompt: "hello".into(),
-            cli_agent_type: "claude-code".into(),
-            profile: None,
-            session_id: None,
-            feature_flags: vec!["myFlag=yes".into()],
-            env: vec![],
-            secret_env: vec![],
-            timeout: 1,
-            active_inputs: vec![],
-        };
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().to_path_buf());
-        let err = run_submit_with_home(args, home).await.unwrap_err();
-        assert!(
-            err.to_string().contains("expected true/false"),
-            "got: {err}"
-        );
-    }
-
-    #[tokio::test]
-    async fn timeout_message_includes_group_and_profile() {
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().to_path_buf());
-        let args = SubmitArgs {
-            group: "test/group".into(),
-            prompt: "hello".into(),
-            cli_agent_type: "claude-code".into(),
-            profile: Some("vm0/large".into()),
-            session_id: None,
-            feature_flags: vec![],
-            env: vec![],
-            secret_env: vec![],
-            timeout: 0,
-            active_inputs: vec![],
-        };
-
-        let err = run_submit_with_home(args, home).await.unwrap_err();
-        let msg = err.to_string();
-        assert!(msg.contains("group: test/group"), "got: {msg}");
-        assert!(msg.contains("profile: vm0/large"), "got: {msg}");
-        assert!(msg.contains("no local runner"), "got: {msg}");
-        assert!(msg.contains("support this profile"), "got: {msg}");
-    }
-
-    #[tokio::test]
-    async fn maximum_timeout_is_accepted() {
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().to_path_buf());
-        let group = "test/group";
-        let group_dir = home.groups_dir().join(group);
-        let mut args = submit_args_for_test();
-        args.group = group.into();
-        args.timeout = MAX_LOCAL_SUBMIT_TIMEOUT_SECS;
-        let watcher = tokio::spawn(wait_for_job_and_write_success(
-            group_dir,
-            crate::profile::DEFAULT_PROFILE.to_owned(),
-        ));
-
-        let code = run_submit_with_home(args, home).await.unwrap();
-        let request = watcher.await.unwrap();
-
-        assert_eq!(code, ExitCode::SUCCESS);
-        assert_eq!(request.prompt, "hello");
-    }
-
-    #[tokio::test]
-    async fn oversized_timeouts_are_rejected_before_publishing_job() {
-        for timeout in [MAX_LOCAL_SUBMIT_TIMEOUT_SECS + 1, u64::MAX] {
-            let dir = tempfile::tempdir().unwrap();
-            let home = HomePaths::with_root(dir.path().to_path_buf());
-            let group = "test/group";
-            let group_dir = home.groups_dir().join(group);
-            let mut args = submit_args_for_test();
-            args.group = group.into();
-            args.timeout = timeout;
-
-            let err = run_submit_with_home(args, home).await.unwrap_err();
-
-            assert!(matches!(&err, RunnerError::Config(_)), "got: {err:?}");
-            let msg = err.to_string();
-            assert!(msg.contains("--timeout"), "got: {msg}");
-            assert!(msg.contains("must be <="), "got: {msg}");
-            assert!(msg.contains(&timeout.to_string()), "got: {msg}");
-            assert!(
-                !group_dir.exists(),
-                "invalid timeout must not create a local queue group directory"
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn timeout_removes_unclaimed_job_from_queue() {
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().to_path_buf());
-        let group = "test/group";
-        let group_dir = home.groups_dir().join(group);
-        let args = SubmitArgs {
-            group: group.into(),
-            prompt: "hello".into(),
-            cli_agent_type: "claude-code".into(),
-            profile: None,
-            session_id: None,
-            feature_flags: vec![],
-            env: vec![],
-            secret_env: vec![],
-            timeout: 0,
-            active_inputs: vec![],
-        };
-
-        let err = run_submit_with_home(args, home).await.unwrap_err();
-
-        let job_dir =
-            local_queue::profile_jobs_dir(&group_dir, crate::profile::DEFAULT_PROFILE).unwrap();
-        let job_files: Vec<_> = std::fs::read_dir(&job_dir)
-            .unwrap()
-            .filter_map(Result::ok)
-            .filter(|entry| entry.path().extension().and_then(|ext| ext.to_str()) == Some("job"))
-            .collect();
-        let result_files: Vec<_> = std::fs::read_dir(local_queue::results_dir(&group_dir))
-            .unwrap()
-            .filter_map(Result::ok)
-            .filter(|entry| entry.path().extension().and_then(|ext| ext.to_str()) == Some("result"))
-            .collect();
-
-        assert!(err.to_string().contains("timeout waiting for local result"));
-        assert!(job_files.is_empty(), "job files left behind: {job_files:?}");
-        assert!(
-            result_files.is_empty(),
-            "result files left behind: {result_files:?}"
-        );
-    }
-}
+mod tests;
